refactor(access): move user_can_access_chat_file to onyx.access.access
Drops the circular-import workaround: the function is authorization
logic, not a raw DB helper, so it belongs alongside get_acl_for_user
and get_access_for_documents. Removes the lazy imports introduced in
the previous commit and updates chat_backend.py to match.

No behavior change.
